fastApi/requirement.txt
anyio==3.6.2
autopep8==2.0.1
azure-core==1.26.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
fastapi==0.89.1
h11==0.14.
httptools==0.5.0
idna==3.4
isodate==0.6.1
mysql-connector-python==8.0.31
protobuf==3.20.1
pycodestyle==2.10.0
pydantic==1.10.4
python-dotenv==0.21.0
PyYAML==6.0
requests==2.28.2
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
watchfiles==0.18.1
websockets==10.4